Created a model and schema folder
