+++ lib/sing_for_needs/model/artist.ex
@@ -1,0 +1,38 @@
+defmodule SingForNeeds.Model.Artist do
+  @moduledoc """
+  The Artists context.
+  """
+
+  import Ecto.Query, warn: false
+  alias SingForNeeds.Repo
+  alias SingForNeeds.Schema.Artist
+
+
+  def list_artists do
+    Repo.all(Artist)
+  end
+
+  def get_artist!(id), do: Repo.get!(Artist, id)
+
+
+  def create_artist(attrs \\ %{}) do
+    %Artist{}
+    |> Artist.changeset(attrs)
+    |> Repo.insert()
+  end
+
+  def update_artist(%Artist{} = artist, attrs) do
+    artist
+    |> Artist.changeset(attrs)
+    |> Repo.update()
+  end
+
+
+  def delete_artist(%Artist{} = artist) do
+    Repo.delete(artist)
+  end
+
+  def change_artist(%Artist{} = artist) do
+    Artist.changeset(artist, %{})
+  end
+end
